to-tfjs/requirements.txt
transformers[tf]
sentence-transformers
tensorflowjs
numpy
